feat: finishes navbar
