+++ package.json
@@ -77,6 +77,7 @@
     "prettier": "^2.8.4",
     "prettier-plugin-tailwindcss": "^0.2.3",
     "storybook": "^7.0.0-beta.61",
+    "storybook-addon-auth0-react": "^1.2.0",
     "storybook-addon-next-router": "^4.0.2",
     "style-loader": "^3.3.1",
     "tailwindcss": "^3.2.6",
