+++ src/components/buttons/Button/Button.stories.tsx
@@ -1,4 +1,4 @@
-import { ChevronRightIcon } from '@heroicons/react/24/outline';
+import { ChevronRightIcon } from '@heroicons/react/24/solid';
 import Button from './Button';
 import { mockButtonProps } from './Button.mocks';
