--- src/components/templates/MainNavbar/MainNavbar.stories.tsx
@@ -1,8 +1,0 @@
-import MainNavbar from './MainNavbar';
-import { mockMainNavbarProps } from './MainNavbar.mocks';
-
-export default { component: MainNavbar };
-
-export const Default = {
-  args: { ...mockMainNavbarProps.base },
-};
