+++ src/pages/privacy-info.tsx
@@ -6,11 +6,7 @@ const PrivacyInfo: NextPageWithLayout = () => {
   return (
     <>
       <div>Hello and welcome to the privacy info page!</div>
-      <Link href="/">
-        <button className="rounded bg-blue-500 py-2 px-4 font-bold text-white hover:bg-blue-700">
-          Go Home
-        </button>
-      </Link>
+      <Link href="/">Home</Link>
     </>
   );
 };
